refactor(ai): migrate prepare_* tools to registry (batch 5)
Move eight prepare tools into registry modules with shared Zod schemas
and prepare-tool-helpers (pending action batch, job draft merge, etc.).
Executor keeps enterprise invite + schedule tools; re-exports
buildPendingEventBatchFromDrafts. Remove unused imports in suggest_mentors.
