+++ src/lib/ai/tools/executor.ts
@@ -1,11 +1,7 @@
 import { z } from "zod";
 import type { SupabaseClient } from "@supabase/supabase-js";
-import type { PrepareEventArgs, ToolName } from "./definitions";
+import type { ToolName } from "./definitions";
 import { TOOL_NAMES } from "./definitions";
-import { getEnterpriseStats } from "./enterprise/stats";
-import { listManagedOrgs } from "./enterprise/managed-orgs";
-import { getEnterpriseOrgCapacity, getEnterpriseQuota } from "./enterprise/quota";
-import { listEnterpriseAuditEvents } from "./enterprise/audit-visibility";
 import { getEnterprisePermissions, type EnterpriseRole } from "@/types/enterprise";
 import { isToolAllowed, type AiActorRole } from "@/lib/ai/access-policy";
 import {
@@ -15,66 +11,21 @@ import {
   withStageTimeout,
 } from "@/lib/ai/timeout";
 import type { AiToolAuthMode } from "@/lib/ai/chat-telemetry";
-import { searchNavigationTargets } from "@/lib/ai/navigation-targets";
-import type { NavConfig } from "@/lib/navigation/nav-items";
-import {
-  assistantAnnouncementDraftSchema,
-  assistantPreparedAnnouncementSchema,
-  type AssistantPreparedAnnouncement,
-} from "@/lib/schemas/content";
-import {
-  assistantJobDraftSchema,
-  assistantPreparedJobSchema,
-  type AssistantPreparedJob,
-} from "@/lib/schemas/jobs";
-import {
-  assistantDiscussionDraftSchema,
-  assistantDiscussionReplyDraftSchema,
-  assistantPreparedDiscussionReplySchema,
-  assistantPreparedDiscussionSchema,
-} from "@/lib/schemas/discussion";
-import {
-  assistantChatMessageDraftSchema,
-  assistantPreparedChatMessageSchema,
-  assistantGroupMessageDraftSchema,
-  assistantPreparedGroupMessageSchema,
-} from "@/lib/schemas/chat-ai";
-import {
-  assistantEventDraftSchema,
-  assistantPreparedEventSchema,
-} from "@/lib/schemas/events-ai";
-import { fetchJobSourceDraft, JobSourceIntakeError } from "@/lib/jobs/source-intake";
 import { ScheduleSecurityError } from "@/lib/schedule-security/errors";
 import { fetchUrlSafe } from "@/lib/schedule-security/fetchUrlSafe";
 import { isOwnedScheduleUploadPath } from "@/lib/ai/schedule-upload-path";
 import {
-  buildPendingActionSummary,
-  type CreateAnnouncementPendingPayload,
-  type SendChatMessagePendingPayload,
-  type SendGroupChatMessagePendingPayload,
-  type CreateDiscussionReplyPendingPayload,
-  createPendingAction,
   createOrRevisePendingAction,
-  type CreateOrReviseResult,
-  type CreateOrReviseFailureReason,
-  type PendingActionPayload,
-  type PendingActionRecord,
-  type CreateDiscussionThreadPendingPayload,
-  type CreateEventPendingPayload,
-  type CreateJobPostingPendingPayload,
   type CreateEnterpriseInvitePendingPayload,
   type RevokeEnterpriseInvitePendingPayload,
 } from "@/lib/ai/pending-actions";
 import { aiLog, type AiLogContext } from "@/lib/ai/logger";
 import { sanitizeIlikeInput } from "@/lib/security/validation";
 import {
-  resolveChatMessageRecipient,
-  type DirectChatSupabase,
-} from "@/lib/chat/direct-chat";
-import {
-  resolveGroupChatTarget,
-  type GroupChatSupabase,
-} from "@/lib/chat/group-chat";
+  buildPendingActionField,
+  createEventPendingActionsFromDrafts,
+  pendingActionFailureToToolError,
+} from "@/lib/ai/tools/prepare-tool-helpers";
 import type {
   ScheduleImageMimeType,
   ScheduleExtractionResult,
@@ -84,17 +35,15 @@ import {
   toolError,
   type ToolExecutionResult,
 } from "@/lib/ai/tools/result";
-import {
-  buildMemberName,
-  getSafeErrorMessage,
-  safeToolQuery,
-} from "@/lib/ai/tools/shared";
+import { getSafeErrorMessage } from "@/lib/ai/tools/shared";
 
 export type {
   ScheduleFileToolErrorCode,
   ToolExecutionErrorCode,
   ToolExecutionResult,
 } from "@/lib/ai/tools/result";
+
+export { buildPendingEventBatchFromDrafts } from "@/lib/ai/tools/prepare-tool-helpers";
 
 export type ToolExecutionAuthorization =
   | {
@@ -126,10 +75,6 @@ export interface ToolExecutionContext {
   };
 }
 
-type CountResult =
-  | { ok: true; count: number }
-  | { ok: false; error: string };
-
 const NON_ADMIN_RLS_READ_TOOL_NAMES: ReadonlySet<ToolName> = new Set<ToolName>([
   "list_announcements",
   "list_events",
@@ -140,108 +85,6 @@ const NON_ADMIN_RLS_READ_TOOL_NAMES: ReadonlySet<ToolName> = new Set<ToolName>([
   "find_navigation_targets",
 ]);
 
-const listDonationsSchema = z
-  .object({
-    limit: z.number().int().min(1).max(25).optional(),
-    status: z.enum(["succeeded", "failed", "pending"]).optional(),
-    purpose: z.string().trim().min(1).optional(),
-  })
-  .strict();
-
-const listParentsSchema = z
-  .object({
-    limit: z.number().int().min(1).max(25).optional(),
-    relationship: z.string().trim().min(1).optional(),
-  })
-  .strict();
-
-const listPhilanthropyEventsSchema = z
-  .object({
-    limit: z.number().int().min(1).max(25).optional(),
-    upcoming: z.boolean().optional(),
-  })
-  .strict();
-
-const prepareAnnouncementSchema = z
-  .object({
-    title: z.string().trim().min(1).optional(),
-    body: z.string().trim().optional(),
-    is_pinned: z.boolean().optional(),
-    audience: z.enum(["all", "members", "active_members", "alumni", "individuals"]).optional(),
-    send_notification: z.boolean().optional(),
-    audience_user_ids: z.array(z.string().uuid()).optional(),
-  })
-  .strict();
-
-const prepareJobPostingSchema = z
-  .object({
-    title: z.string().trim().min(1).optional(),
-    company: z.string().trim().min(1).optional(),
-    location: z.string().trim().min(1).optional(),
-    location_type: z.enum(["remote", "hybrid", "onsite"]).optional(),
-    description: z.string().trim().min(1).optional(),
-    application_url: z.string().trim().min(1).optional(),
-    contact_email: z.string().trim().min(1).optional(),
-    industry: z.string().trim().min(1).optional(),
-    experience_level: z.enum(["entry", "mid", "senior", "lead", "executive"]).optional(),
-    expires_at: z.string().datetime().optional().nullable(),
-    mediaIds: z.array(z.string().uuid()).optional(),
-  })
-  .strict();
-
-const prepareDiscussionThreadSchema = z
-  .object({
-    title: z.string().trim().min(1).optional(),
-    body: z.string().trim().min(1).optional(),
-    mediaIds: z.array(z.string().uuid()).optional(),
-  })
-  .strict();
-
-const prepareDiscussionReplySchema = z
-  .object({
-    discussion_thread_id: z.string().uuid().optional(),
-    thread_title: z.string().trim().min(1).optional(),
-    body: z.string().trim().min(1).optional(),
-  })
-  .strict();
-
-const prepareChatMessageSchema = z
-  .object({
-    recipient_member_id: z.string().uuid().optional(),
-    person_query: z.string().trim().min(1).optional(),
-    body: z.string().trim().min(1).optional(),
-  })
-  .strict();
-
-const prepareGroupMessageSchema = z
-  .object({
-    chat_group_id: z.string().uuid().optional(),
-    group_name_query: z.string().trim().min(1).optional(),
-    body: z.string().trim().min(1).optional(),
-  })
-  .strict();
-
-const prepareEventSchema = z
-  .object({
-    title: z.string().trim().optional(),
-    description: z.string().trim().optional(),
-    start_date: z.string().trim().optional(),
-    start_time: z.string().trim().optional(),
-    end_date: z.string().trim().optional(),
-    end_time: z.string().trim().optional(),
-    location: z.string().trim().optional(),
-    event_type: z
-      .enum(["general", "philanthropy", "game", "practice", "meeting", "social", "workout", "fundraiser", "class"])
-      .optional(),
-    is_philanthropy: z.boolean().optional(),
-  })
-  .strict();
-
-const prepareEventsBatchSchema = z
-  .object({
-    events: z.array(prepareEventSchema).min(1).max(10),
-  })
-  .strict();
 const scrapeScheduleWebsiteSchema = z
   .object({
     url: z.string().url(),
@@ -249,24 +92,6 @@ const scrapeScheduleWebsiteSchema = z
   .strict();
 const extractSchedulePdfSchema = z.object({}).strict();
 
-const getOrgStatsSchema = z.object({}).strict();
-const getDonationAnalyticsSchema = z
-  .object({
-    window_days: z.number().int().min(7).max(3650).optional(),
-    bucket: z.enum(["day", "week", "month"]).optional(),
-    top_purposes_limit: z.number().int().min(1).max(10).optional(),
-  })
-  .strict();
-const getEnterpriseStatsSchema = z.object({}).strict();
-const getEnterpriseQuotaSchema = z.object({}).strict();
-const getEnterpriseOrgCapacitySchema = z.object({}).strict();
-const listManagedOrgsSchema = z.object({}).strict();
-const listEnterpriseAuditEventsSchema = z
-  .object({
-    organization_id: z.string().trim().min(1).optional(),
-    limit: z.number().int().min(1).max(100).optional(),
-  })
-  .strict();
 const prepareEnterpriseInviteSchema = z
   .object({
     role: z.enum(["admin", "active_member", "alumni"]).optional(),
@@ -288,77 +113,11 @@ const revokeEnterpriseInviteSchema = z
       (typeof value.invite_code === "string" && value.invite_code.length > 0),
     { message: "Expected invite_id or invite_code" },
   );
-const suggestConnectionsSchema = z
-  .object({
-    person_type: z.enum(["member", "alumni"]).optional(),
-    person_id: z.string().uuid().optional(),
-    person_query: z.string().trim().min(1).optional(),
-    limit: z.number().int().min(1).max(25).optional(),
-  })
-  .refine(
-    (value) =>
-      (typeof value.person_query === "string" && value.person_query.length > 0) ||
-      (typeof value.person_type === "string" && typeof value.person_id === "string"),
-    {
-      message: "Expected person_query or both person_type and person_id",
-    }
-  )
-  .strict();
-const suggestMentorsSchema = z
-  .object({
-    mentee_id: z.string().uuid().optional(),
-    mentee_query: z.string().trim().min(1).optional(),
-    focus_areas: z.array(z.string().trim().min(1)).optional(),
-    limit: z.number().int().min(1).max(25).optional(),
-  })
-  .refine(
-    (value) =>
-      (typeof value.mentee_query === "string" && value.mentee_query.length > 0) ||
-      typeof value.mentee_id === "string",
-    {
-      message: "Expected mentee_query or mentee_id",
-    }
-  )
-  .strict();
-const listAvailableMentorsSchema = z
-  .object({
-    limit: z.number().int().min(1).max(25).optional(),
-  })
-  .strict();
-const findNavigationTargetsSchema = z
-  .object({
-    query: z.string().trim().min(1),
-    limit: z.number().int().min(1).max(10).optional(),
-  })
-  .strict();
-
 const ARG_SCHEMAS: Partial<Record<ToolName, z.ZodSchema>> = {
-  list_donations: listDonationsSchema,
-  list_parents: listParentsSchema,
-  list_philanthropy_events: listPhilanthropyEventsSchema,
-  list_managed_orgs: listManagedOrgsSchema,
-  list_enterprise_audit_events: listEnterpriseAuditEventsSchema,
   prepare_enterprise_invite: prepareEnterpriseInviteSchema,
   revoke_enterprise_invite: revokeEnterpriseInviteSchema,
-  prepare_announcement: prepareAnnouncementSchema,
-  prepare_job_posting: prepareJobPostingSchema,
-  prepare_chat_message: prepareChatMessageSchema,
-  prepare_group_message: prepareGroupMessageSchema,
-  prepare_discussion_reply: prepareDiscussionReplySchema,
-  prepare_discussion_thread: prepareDiscussionThreadSchema,
-  prepare_event: prepareEventSchema,
-  prepare_events_batch: prepareEventsBatchSchema,
   scrape_schedule_website: scrapeScheduleWebsiteSchema,
   extract_schedule_pdf: extractSchedulePdfSchema,
-  get_org_stats: getOrgStatsSchema,
-  get_donation_analytics: getDonationAnalyticsSchema,
-  get_enterprise_stats: getEnterpriseStatsSchema,
-  get_enterprise_quota: getEnterpriseQuotaSchema,
-  get_enterprise_org_capacity: getEnterpriseOrgCapacitySchema,
-  suggest_connections: suggestConnectionsSchema,
-  suggest_mentors: suggestMentorsSchema,
-  list_available_mentors: listAvailableMentorsSchema,
-  find_navigation_targets: findNavigationTargetsSchema,
 };
 
 const ENTERPRISE_TOOL_NAMES = new Set<ToolName>([
@@ -402,57 +161,6 @@ function validateArgs(
   return { valid: true, args: parsed.data };
 }
 
-function buildPendingActionField(
-  created: Extract<CreateOrReviseResult, { record: PendingActionRecord }>,
-  payload: PendingActionPayload
-) {
-  return {
-    id: created.record.id,
-    action_type: created.record.action_type,
-    payload,
-    expires_at: created.record.expires_at,
-    summary: buildPendingActionSummary(created.record),
-    ...(created.revised
-      ? {
-          revise_count: created.reviseCount,
-          previous_payload: created.previousPayload,
-        }
-      : {}),
-  };
-}
-
-function pendingActionFailureToToolError(
-  reason: CreateOrReviseFailureReason
-): ToolExecutionResult {
-  switch (reason) {
-    case "revise_limit":
-      return toolError(
-        "Maximum revisions reached for this draft. Please confirm or cancel it before making more changes.",
-        "pending_action_revise_limit"
-      );
-    case "not_pending":
-      return toolError(
-        "That draft is no longer open for edits. Start a new request to create another draft.",
-        "pending_action_not_pending"
-      );
-    case "not_found":
-      return toolError(
-        "That draft is no longer available. Start a new request to create another draft.",
-        "pending_action_not_pending"
-      );
-    case "action_type_mismatch":
-      return toolError(
-        "That draft is for a different action type. Please confirm or cancel it before starting a new one.",
-        "pending_action_not_pending"
-      );
-    case "conflict":
-      return toolError(
-        "Another revision landed first — please retry your edit.",
-        "pending_action_conflict"
-      );
-  }
-}
-
 function isScheduleImageConfigurationError(error: unknown): boolean {
   const message = getSafeErrorMessage(error);
   return /ZAI_IMAGE_MODEL|vision model|model such as glm-5v-turbo/i.test(message);
@@ -467,36 +175,6 @@ function buildLogContext(
     userId: ctx.userId,
     ...(ctx.threadId ? { threadId: ctx.threadId } : {}),
   };
-}
-
-async function safeToolCount(
-  logContext: AiLogContext,
-  fn: () => Promise<{ count: number | null; error: unknown }>
-): Promise<CountResult> {
-  try {
-    const { count, error } = await fn();
-    if (error || count === null) {
-      if (error) {
-        aiLog("warn", "ai-tools", "count query failed", logContext, {
-          error: getSafeErrorMessage(error),
-        });
-      } else {
-        aiLog("warn", "ai-tools", "count query failed", logContext, {
-          error: "count_unavailable",
-        });
-      }
-      return { ok: false, error: "Query failed" };
-    }
-    return { ok: true, count };
-  } catch (err) {
-    if (isStageTimeoutError(err)) {
-      throw err;
-    }
-    aiLog("warn", "ai-tools", "unexpected count error", logContext, {
-      error: getSafeErrorMessage(err),
-    });
-    return { ok: false, error: "Unexpected error" };
-  }
 }
 
 const SCRAPE_SCHEDULE_FETCH_TIMEOUT_MS = 10_000;
@@ -564,569 +242,10 @@ interface MembershipRow {
   status: string | null;
 }
 
-interface EventPendingActionRecord {
-  id: string;
-  action_type: string;
-  payload: CreateEventPendingPayload;
-  expires_at: string;
-  summary: { title: string; description: string };
-}
-
-interface EventValidationErrorRecord {
-  index: number;
-  missing_fields: string[];
-  draft: Record<string, unknown>;
-}
-
-async function listDonations(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof listDonationsSchema>,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const limit = Math.min(args.limit ?? 10, 25);
-  return safeToolQuery(logContext, async () => {
-    let query = sb
-      .from("organization_donations")
-      .select("id, donor_name, donor_email, amount_cents, purpose, status, created_at, anonymous")
-      .eq("organization_id", orgId)
-      .is("deleted_at", null)
-      .order("created_at", { ascending: false })
-      .limit(limit);
-
-    if (args.status) {
-      query = query.eq("status", args.status);
-    }
-    if (args.purpose) {
-      query = query.ilike("purpose", `%${sanitizeIlikeInput(args.purpose)}%`);
-    }
-
-    const { data, error } = await query;
-
-    if (!Array.isArray(data) || error) {
-      return { data, error };
-    }
-
-    return {
-      data: data.map((row) => {
-        const isAnonymous = Boolean(row.anonymous);
-        return {
-          id: row.id,
-          donor_name: isAnonymous ? "Anonymous" : (row.donor_name ?? null),
-          donor_email: isAnonymous ? "Anonymous" : (row.donor_email ?? null),
-          amount_dollars: typeof row.amount_cents === "number" ? row.amount_cents / 100 : null,
-          purpose: row.purpose ?? null,
-          status: row.status ?? null,
-          created_at: row.created_at ?? null,
-          anonymous: isAnonymous,
-        };
-      }),
-      error: null,
-    };
-  });
-}
-
-async function listParents(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof listParentsSchema>,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const limit = Math.min(args.limit ?? 10, 25);
-  return safeToolQuery(logContext, async () => {
-    let query = sb
-      .from("parents")
-      .select("id, first_name, last_name, email, relationship, student_name, phone_number")
-      .eq("organization_id", orgId)
-      .is("deleted_at", null)
-      .order("last_name", { ascending: true })
-      .limit(limit);
-
-    if (args.relationship) {
-      query = query.ilike("relationship", `%${sanitizeIlikeInput(args.relationship)}%`);
-    }
-
-    const { data, error } = await query;
-
-    if (!Array.isArray(data) || error) {
-      return { data, error };
-    }
-
-    return {
-      data: data.map((row) => ({
-        id: row.id,
-        name: buildMemberName(row.first_name ?? "", row.last_name ?? ""),
-        relationship: row.relationship ?? null,
-        student_name: row.student_name ?? null,
-        email: row.email ?? null,
-        phone_number: row.phone_number ?? null,
-      })),
-      error: null,
-    };
-  });
-}
-
-async function listPhilanthropyEvents(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof listPhilanthropyEventsSchema>,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const limit = Math.min(args.limit ?? 10, 25);
-  const upcoming = args.upcoming ?? true;
-  const now = new Date().toISOString();
-  return safeToolQuery(logContext, () => {
-    let query = sb
-      .from("events")
-      .select("id, title, start_date, end_date, location, description")
-      .eq("organization_id", orgId)
-      .is("deleted_at", null)
-      .or("is_philanthropy.eq.true,event_type.eq.philanthropy")
-      .order("start_date", { ascending: upcoming })
-      .limit(limit);
-    if (upcoming) {
-      query = query.gte("start_date", now);
-    } else {
-      query = query.lt("start_date", now);
-    }
-    return query;
-  });
-}
-
-const REQUIRED_PREPARED_JOB_FIELDS: Array<keyof AssistantPreparedJob> = [
-  "title",
-  "company",
-  "location",
-  "industry",
-  "experience_level",
-  "description",
-];
-const REQUIRED_PREPARED_ANNOUNCEMENT_FIELDS: Array<keyof AssistantPreparedAnnouncement> = [
-  "title",
-];
-const REQUIRED_PREPARED_EVENT_FIELDS = [
-  "title",
-  "start_date",
-  "start_time",
-] as const satisfies ReadonlyArray<keyof z.infer<typeof assistantPreparedEventSchema>>;
-
-async function createEventPendingActionsFromDrafts(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  events: PrepareEventArgs[],
-  logContext: AiLogContext,
-  orgSlug: string | null
-): Promise<{
-  pendingActions: EventPendingActionRecord[];
-  validationErrors: EventValidationErrorRecord[];
-}> {
-  const threadId = ctx.threadId;
-  if (!threadId) {
-    throw new Error("Event preparation requires a thread context");
-  }
-
-  const pendingActions: EventPendingActionRecord[] = [];
-  const validationErrors: EventValidationErrorRecord[] = [];
-
-  for (let i = 0; i < events.length; i++) {
-    const eventArgs = events[i];
-    const normalized = Object.fromEntries(
-      Object.entries(eventArgs).filter(
-        ([, value]) => !(typeof value === "string" && value.trim().length === 0)
-      )
-    ) as PrepareEventArgs;
-
-    const draftWithDefaults = {
-      ...normalized,
-      event_type: normalized.event_type ?? "general",
-      is_philanthropy:
-        normalized.is_philanthropy ?? normalized.event_type === "philanthropy",
-    };
-
-    const parsedDraft = assistantEventDraftSchema.safeParse(draftWithDefaults);
-    if (!parsedDraft.success) {
-      validationErrors.push({
-        index: i,
-        missing_fields: parsedDraft.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: draftWithDefaults,
-      });
-      continue;
-    }
-
-    const missingFields = REQUIRED_PREPARED_EVENT_FIELDS.filter((field) => {
-      const value = parsedDraft.data[field];
-      return typeof value !== "string" || value.trim().length === 0;
-    });
-
-    if (missingFields.length > 0) {
-      validationErrors.push({
-        index: i,
-        missing_fields: [...missingFields],
-        draft: parsedDraft.data as unknown as Record<string, unknown>,
-      });
-      continue;
-    }
-
-    const prepared = assistantPreparedEventSchema.safeParse(parsedDraft.data);
-    if (!prepared.success) {
-      validationErrors.push({
-        index: i,
-        missing_fields: prepared.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: parsedDraft.data as unknown as Record<string, unknown>,
-      });
-      continue;
-    }
-
-    const pendingPayload: CreateEventPendingPayload = {
-      ...prepared.data,
-      orgSlug,
-    };
-
-    try {
-      const pendingAction = await createPendingAction(sb, {
-        organizationId: ctx.orgId,
-        userId: ctx.userId,
-        threadId,
-        actionType: "create_event",
-        payload: pendingPayload,
-      });
-      const summary = buildPendingActionSummary(pendingAction);
-      pendingActions.push({
-        id: pendingAction.id,
-        action_type: pendingAction.action_type,
-        payload: pendingPayload,
-        expires_at: pendingAction.expires_at,
-        summary,
-      });
-    } catch (error) {
-      aiLog("warn", "ai-tools", "event pending action insert failed", logContext, {
-        index: i,
-        error: getSafeErrorMessage(error),
-      });
-      validationErrors.push({
-        index: i,
-        missing_fields: ["_insert_failed"],
-        draft: prepared.data as unknown as Record<string, unknown>,
-      });
-    }
-  }
-
-  return {
-    pendingActions,
-    validationErrors,
-  };
-}
-
-export async function buildPendingEventBatchFromDrafts(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  events: PrepareEventArgs[],
-  logContext: AiLogContext,
-  orgSlug: string | null
-): Promise<{
-  state: "missing_fields" | "needs_batch_confirmation";
-  pending_actions?: Array<{
-    id: string;
-    action_type: string;
-    payload: CreateEventPendingPayload;
-    expires_at: string;
-    summary: { title: string; description: string };
-  }>;
-  validation_errors?: Array<{
-    index: number;
-    missing_fields: string[];
-    draft: Record<string, unknown>;
-  }>;
-}> {
-  const { pendingActions, validationErrors } = await createEventPendingActionsFromDrafts(
-    sb,
-    ctx,
-    events,
-    logContext,
-    orgSlug
-  );
-
-  if (pendingActions.length === 0) {
-    return {
-      state: "missing_fields",
-      validation_errors: validationErrors,
-    };
-  }
-
-  return {
-    state: "needs_batch_confirmation",
-    pending_actions: pendingActions,
-    validation_errors: validationErrors.length > 0 ? validationErrors : undefined,
-  };
-}
-
-function sanitizeDraftValue(value: unknown): string | undefined {
-  if (typeof value !== "string") {
-    return undefined;
-  }
-  const normalized = value.trim();
-  return normalized.length > 0 ? normalized : undefined;
-}
-
-function normalizeAssistantDraft(
-  args: z.infer<typeof prepareJobPostingSchema>
-): z.infer<typeof prepareJobPostingSchema> {
-  const stringFields = [
-    "title", "company", "location", "description",
-    "application_url", "contact_email", "industry",
-  ] as const;
-
-  const overrides: Record<string, string> = {};
-  for (const field of stringFields) {
-    const sanitized = sanitizeDraftValue(args[field]);
-    if (sanitized) {
-      overrides[field] = sanitized;
-    }
-  }
-
-  return { ...args, ...overrides };
-}
-
-function mergeDrafts<T extends Record<string, unknown>>(primary: T, secondary: Partial<T>): T {
-  const merged = { ...secondary, ...primary };
-  return Object.fromEntries(
-    Object.entries(merged).filter(
-      ([, value]) => !(typeof value === "string" && value.trim().length === 0)
-    )
-  ) as T;
-}
-
-function hasPreparedJobRequirements(
-  draft: Partial<z.infer<typeof prepareJobPostingSchema>>
-): boolean {
-  const hasRequiredFields = REQUIRED_PREPARED_JOB_FIELDS.every((field) => {
-    const value = draft[field];
-    return typeof value === "string" && value.trim().length > 0;
-  });
-
-  if (!hasRequiredFields) {
-    return false;
-  }
-
-  const hasApplicationUrl =
-    typeof draft.application_url === "string" && draft.application_url.trim().length > 0;
-  const hasContactEmail =
-    typeof draft.contact_email === "string" && draft.contact_email.trim().length > 0;
-
-  return hasApplicationUrl || hasContactEmail;
-}
-
-async function prepareJobPosting(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareJobPostingSchema>
-): Promise<ToolExecutionResult> {
-  const logContext = buildLogContext(ctx);
-  if (!ctx.threadId) {
-    return toolError("Job preparation requires a thread context");
-  }
-
-  const parsedDraft = assistantJobDraftSchema.safeParse(normalizeAssistantDraft(args));
-  if (!parsedDraft.success) {
-    return toolError("Invalid job draft");
-  }
-
-  let sourceDraft: Partial<z.infer<typeof prepareJobPostingSchema>> = {};
-  let sourceWarning: string | null = null;
-  if (parsedDraft.data.application_url && !hasPreparedJobRequirements(parsedDraft.data)) {
-    try {
-      sourceDraft = await fetchJobSourceDraft(parsedDraft.data.application_url);
-    } catch (error) {
-      if (error instanceof JobSourceIntakeError) {
-        sourceWarning = error.message;
-      }
-      if (!(error instanceof JobSourceIntakeError)) {
-        return toolError("Unable to read the job posting URL");
-      }
-    }
-  }
-
-  const mergedDraft = mergeDrafts(parsedDraft.data, sourceDraft);
-  const missingFields = REQUIRED_PREPARED_JOB_FIELDS.filter((field) => {
-    const value = mergedDraft[field];
-    return typeof value !== "string" || value.trim().length === 0;
-  });
-
-  const hasApplicationUrl =
-    typeof mergedDraft.application_url === "string" && mergedDraft.application_url.trim().length > 0;
-  const hasContactEmail =
-    typeof mergedDraft.contact_email === "string" && mergedDraft.contact_email.trim().length > 0;
-  if (!hasApplicationUrl && !hasContactEmail) {
-    missingFields.push("application_url");
-  }
-
-  if (missingFields.length > 0) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: Array.from(new Set(missingFields)),
-        draft: mergedDraft,
-        sourced_fields: Object.keys(sourceDraft),
-        ...(sourceWarning ? { source_warning: sourceWarning } : {}),
-      },
-    };
-  }
-
-  const prepared = assistantPreparedJobSchema.safeParse(mergedDraft);
-  if (!prepared.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: prepared.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: mergedDraft,
-        sourced_fields: Object.keys(sourceDraft),
-        ...(sourceWarning ? { source_warning: sourceWarning } : {}),
-      },
-    };
-  }
-
-  const { data: org, error: orgError } = await sb
-    .from("organizations")
-    .select("slug")
-    .eq("id", ctx.orgId)
-    .maybeSingle();
-
-  if (orgError) {
-    aiLog("warn", "ai-tools", "prepare_job_posting org lookup failed", logContext, {
-      error: getSafeErrorMessage(orgError),
-    });
-    return toolError("Failed to load organization context");
-  }
-
-  const pendingPayload: CreateJobPostingPendingPayload = {
-    ...prepared.data,
-    orgSlug: typeof org?.slug === "string" ? org.slug : null,
-  };
-  const created = await createOrRevisePendingAction(sb, {
-    organizationId: ctx.orgId,
-    userId: ctx.userId,
-    threadId: ctx.threadId,
-    actionType: "create_job_posting",
-    payload: pendingPayload,
-    activeActionId: ctx.activePendingActionId,
-  });
-  if ("failed" in created) return pendingActionFailureToToolError(created.reason);
-  return {
-    kind: "ok",
-    data: {
-      state: "needs_confirmation",
-      draft: prepared.data,
-      pending_action: buildPendingActionField(created, pendingPayload),
-      sourced_fields: Object.keys(sourceDraft),
-    },
-  };
-}
-
-async function prepareAnnouncement(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareAnnouncementSchema>
-): Promise<ToolExecutionResult> {
-  const logContext = buildLogContext(ctx);
-  if (!ctx.threadId) {
-    return toolError("Announcement preparation requires a thread context");
-  }
-
-  const normalizedDraft = {
-    ...Object.fromEntries(
-      Object.entries({
-        ...args,
-        title: sanitizeDraftValue(args.title),
-        body: sanitizeDraftValue(args.body),
-      }).filter(([, value]) => value !== undefined)
-    ),
-    audience: args.audience ?? "all",
-    is_pinned: args.is_pinned ?? false,
-    send_notification: args.send_notification ?? false,
-  };
-
-  const parsedDraft = assistantAnnouncementDraftSchema.safeParse(normalizedDraft);
-  if (!parsedDraft.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: parsedDraft.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: normalizedDraft,
-      },
-    };
-  }
-
-  const missingFields = REQUIRED_PREPARED_ANNOUNCEMENT_FIELDS.filter((field) => {
-    const value = parsedDraft.data[field];
-    return typeof value !== "string" || value.trim().length === 0;
-  });
-
-  if (missingFields.length > 0) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: missingFields,
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const prepared = assistantPreparedAnnouncementSchema.safeParse(parsedDraft.data);
-  if (!prepared.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: prepared.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const { data: org, error: orgError } = await sb
-    .from("organizations")
-    .select("slug")
-    .eq("id", ctx.orgId)
-    .maybeSingle();
-
-  if (orgError) {
-    aiLog("warn", "ai-tools", "prepare_announcement org lookup failed", logContext, {
-      error: getSafeErrorMessage(orgError),
-    });
-    return toolError("Failed to load organization context");
-  }
-
-  const pendingPayload: CreateAnnouncementPendingPayload = {
-    ...prepared.data,
-    orgSlug: typeof org?.slug === "string" ? org.slug : null,
-  };
-  const created = await createOrRevisePendingAction(sb, {
-    organizationId: ctx.orgId,
-    userId: ctx.userId,
-    threadId: ctx.threadId,
-    actionType: "create_announcement",
-    payload: pendingPayload,
-    activeActionId: ctx.activePendingActionId,
-  });
-  if ("failed" in created) return pendingActionFailureToToolError(created.reason);
-  return {
-    kind: "ok",
-    data: {
-      state: "needs_confirmation",
-      draft: prepared.data,
-      pending_action: buildPendingActionField(created, pendingPayload),
-    },
-  };
-}
-
 async function prepareEnterpriseInvite(
   sb: SB,
   ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareEnterpriseInviteSchema>,
+  args: z.infer<typeof prepareEnterpriseInviteSchema>
 ): Promise<ToolExecutionResult> {
   const logContext = buildLogContext(ctx);
   if (!ctx.threadId) {
@@ -1251,7 +370,7 @@ async function prepareEnterpriseInvite(
 async function revokeEnterpriseInvite(
   sb: SB,
   ctx: ToolExecutionContext,
-  args: z.infer<typeof revokeEnterpriseInviteSchema>,
+  args: z.infer<typeof revokeEnterpriseInviteSchema>
 ): Promise<ToolExecutionResult> {
   if (!ctx.threadId) {
     return toolError("Enterprise invite revocation requires a thread context");
@@ -1320,632 +439,6 @@ async function revokeEnterpriseInvite(
       draft: pendingPayload,
       pending_action: buildPendingActionField(created, pendingPayload),
     },
-  };
-}
-
-async function prepareDiscussionThread(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareDiscussionThreadSchema>
-): Promise<ToolExecutionResult> {
-  const logContext = buildLogContext(ctx);
-  if (!ctx.threadId) {
-    return toolError("Discussion preparation requires a thread context");
-  }
-
-  const parsedDraft = assistantDiscussionDraftSchema.safeParse(args);
-  if (!parsedDraft.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: parsedDraft.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: args,
-      },
-    };
-  }
-
-  const missingFields: string[] = [];
-  if (!parsedDraft.data.title) missingFields.push("title");
-  if (!parsedDraft.data.body) missingFields.push("body");
-
-  if (missingFields.length > 0) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: missingFields,
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const prepared = assistantPreparedDiscussionSchema.safeParse(parsedDraft.data);
-  if (!prepared.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: prepared.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const { data: org, error: orgError } = await sb
-    .from("organizations")
-    .select("slug")
-    .eq("id", ctx.orgId)
-    .maybeSingle();
-
-  if (orgError) {
-    aiLog("warn", "ai-tools", "prepare_discussion_thread org lookup failed", logContext, {
-      error: getSafeErrorMessage(orgError),
-    });
-    return toolError("Failed to load organization context");
-  }
-
-  const pendingPayload: CreateDiscussionThreadPendingPayload = {
-    ...prepared.data,
-    orgSlug: typeof org?.slug === "string" ? org.slug : null,
-  };
-  const created = await createOrRevisePendingAction(sb, {
-    organizationId: ctx.orgId,
-    userId: ctx.userId,
-    threadId: ctx.threadId,
-    actionType: "create_discussion_thread",
-    payload: pendingPayload,
-    activeActionId: ctx.activePendingActionId,
-  });
-  if ("failed" in created) return pendingActionFailureToToolError(created.reason);
-  return {
-    kind: "ok",
-    data: {
-      state: "needs_confirmation",
-      draft: prepared.data,
-      pending_action: buildPendingActionField(created, pendingPayload),
-    },
-  };
-}
-
-async function prepareDiscussionReply(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareDiscussionReplySchema>
-): Promise<ToolExecutionResult> {
-  const logContext = buildLogContext(ctx);
-  if (!ctx.threadId) {
-    return toolError("Discussion reply preparation requires a thread context");
-  }
-
-  const normalizedDraft = {
-    ...Object.fromEntries(
-      Object.entries({
-        ...args,
-        thread_title: sanitizeDraftValue(args.thread_title),
-        body: sanitizeDraftValue(args.body),
-      }).filter(([, value]) => value !== undefined)
-    ),
-  };
-  const parsedDraft = assistantDiscussionReplyDraftSchema.safeParse(normalizedDraft);
-  if (!parsedDraft.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: parsedDraft.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: normalizedDraft,
-      },
-    };
-  }
-
-  const missingFields: string[] = [];
-  if (!parsedDraft.data.discussion_thread_id) missingFields.push("discussion_thread_id");
-  if (!parsedDraft.data.body) missingFields.push("body");
-
-  if (missingFields.length > 0) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: missingFields,
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const prepared = assistantPreparedDiscussionReplySchema.safeParse(parsedDraft.data);
-  if (!prepared.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: prepared.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const { data: org, error: orgError } = await sb
-    .from("organizations")
-    .select("slug")
-    .eq("id", ctx.orgId)
-    .maybeSingle();
-
-  if (orgError) {
-    aiLog("warn", "ai-tools", "prepare_discussion_reply org lookup failed", logContext, {
-      error: getSafeErrorMessage(orgError),
-    });
-    return toolError("Failed to load organization context");
-  }
-
-  const pendingPayload: CreateDiscussionReplyPendingPayload = {
-    ...prepared.data,
-    orgSlug: typeof org?.slug === "string" ? org.slug : null,
-  };
-  const created = await createOrRevisePendingAction(sb, {
-    organizationId: ctx.orgId,
-    userId: ctx.userId,
-    threadId: ctx.threadId,
-    actionType: "create_discussion_reply",
-    payload: pendingPayload,
-    activeActionId: ctx.activePendingActionId,
-  });
-  if ("failed" in created) return pendingActionFailureToToolError(created.reason);
-  return {
-    kind: "ok",
-    data: {
-      state: "needs_confirmation",
-      draft: prepared.data,
-      pending_action: buildPendingActionField(created, pendingPayload),
-    },
-  };
-}
-
-async function prepareChatMessage(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareChatMessageSchema>
-): Promise<ToolExecutionResult> {
-  const logContext = buildLogContext(ctx);
-  if (!ctx.threadId) {
-    return toolError("Chat message preparation requires a thread context");
-  }
-
-  const normalizedDraft = {
-    ...Object.fromEntries(
-      Object.entries({
-        recipient_member_id: args.recipient_member_id,
-        person_query: sanitizeDraftValue(args.person_query),
-        body: sanitizeDraftValue(args.body),
-      }).filter(([, value]) => value !== undefined)
-    ),
-  };
-
-  const parsedDraft = assistantChatMessageDraftSchema.safeParse(normalizedDraft);
-  if (!parsedDraft.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: parsedDraft.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: normalizedDraft,
-      },
-    };
-  }
-
-  const recipientResolution = await resolveChatMessageRecipient(sb as DirectChatSupabase, {
-    organizationId: ctx.orgId,
-    senderUserId: ctx.userId,
-    recipientMemberId: parsedDraft.data.recipient_member_id,
-    personQuery: parsedDraft.data.person_query,
-  });
-
-  if (recipientResolution.kind === "recipient_required") {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: [
-          ...(parsedDraft.data.body ? [] : ["body"]),
-          "person_query",
-        ],
-        clarification_kind: "recipient_required",
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  if (recipientResolution.kind === "ambiguous") {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: [
-          ...(parsedDraft.data.body ? [] : ["body"]),
-          "person_query",
-        ],
-        clarification_kind: "recipient_ambiguous",
-        requested_recipient: recipientResolution.requestedRecipient,
-        candidate_recipients: recipientResolution.candidateRecipients,
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  if (recipientResolution.kind === "unavailable") {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: parsedDraft.data.body ? [] : ["body"],
-        clarification_kind: "recipient_unavailable",
-        requested_recipient: recipientResolution.requestedRecipient ?? null,
-        unavailable_reason: recipientResolution.reason,
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const draftWithResolvedRecipient = {
-    ...parsedDraft.data,
-    recipient_member_id: recipientResolution.memberId,
-  };
-
-  if (!parsedDraft.data.body) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: ["body"],
-        draft: draftWithResolvedRecipient,
-      },
-    };
-  }
-
-  const prepared = assistantPreparedChatMessageSchema.safeParse({
-    recipient_member_id: recipientResolution.memberId,
-    recipient_user_id: recipientResolution.userId,
-    recipient_display_name: recipientResolution.displayName,
-    body: parsedDraft.data.body,
-    existing_chat_group_id: recipientResolution.existingChatGroupId ?? undefined,
-  });
-
-  if (!prepared.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: prepared.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: draftWithResolvedRecipient,
-      },
-    };
-  }
-
-  const { data: org, error: orgError } = await sb
-    .from("organizations")
-    .select("slug")
-    .eq("id", ctx.orgId)
-    .maybeSingle();
-
-  if (orgError) {
-    aiLog("warn", "ai-tools", "prepare_chat_message org lookup failed", logContext, {
-      error: getSafeErrorMessage(orgError),
-    });
-    return toolError("Failed to load organization context");
-  }
-
-  const pendingPayload: SendChatMessagePendingPayload = {
-    ...prepared.data,
-    orgSlug: typeof org?.slug === "string" ? org.slug : null,
-  };
-  const created = await createOrRevisePendingAction(sb, {
-    organizationId: ctx.orgId,
-    userId: ctx.userId,
-    threadId: ctx.threadId,
-    actionType: "send_chat_message",
-    payload: pendingPayload,
-    activeActionId: ctx.activePendingActionId,
-  });
-  if ("failed" in created) return pendingActionFailureToToolError(created.reason);
-  return {
-    kind: "ok",
-    data: {
-      state: "needs_confirmation",
-      draft: draftWithResolvedRecipient,
-      pending_action: buildPendingActionField(created, pendingPayload),
-    },
-  };
-}
-
-async function prepareGroupMessage(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareGroupMessageSchema>
-): Promise<ToolExecutionResult> {
-  const logContext = buildLogContext(ctx);
-  if (!ctx.threadId) {
-    return toolError("Group message preparation requires a thread context");
-  }
-
-  const normalizedDraft = {
-    ...Object.fromEntries(
-      Object.entries({
-        chat_group_id: args.chat_group_id,
-        group_name_query: sanitizeDraftValue(args.group_name_query),
-        body: sanitizeDraftValue(args.body),
-      }).filter(([, value]) => value !== undefined)
-    ),
-  };
-
-  const parsedDraft = assistantGroupMessageDraftSchema.safeParse(normalizedDraft);
-  if (!parsedDraft.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: parsedDraft.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: normalizedDraft,
-      },
-    };
-  }
-
-  const groupResolution = await resolveGroupChatTarget(sb as GroupChatSupabase, {
-    organizationId: ctx.orgId,
-    userId: ctx.userId,
-    chatGroupId: parsedDraft.data.chat_group_id,
-    groupNameQuery: parsedDraft.data.group_name_query,
-  });
-
-  if (groupResolution.kind === "group_required") {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: [
-          ...(parsedDraft.data.body ? [] : ["body"]),
-          "group_name_query",
-        ],
-        clarification_kind: "group_required",
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  if (groupResolution.kind === "ambiguous") {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: [
-          ...(parsedDraft.data.body ? [] : ["body"]),
-          "group_name_query",
-        ],
-        clarification_kind: "group_ambiguous",
-        requested_group: groupResolution.requestedGroup,
-        candidate_groups: groupResolution.candidateGroups,
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  if (groupResolution.kind === "unavailable") {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: parsedDraft.data.body ? [] : ["body"],
-        clarification_kind: "group_unavailable",
-        requested_group: groupResolution.requestedGroup ?? null,
-        unavailable_reason: groupResolution.reason,
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const draftWithResolvedGroup = {
-    ...parsedDraft.data,
-    chat_group_id: groupResolution.chatGroupId,
-  };
-
-  if (!parsedDraft.data.body) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: ["body"],
-        draft: draftWithResolvedGroup,
-      },
-    };
-  }
-
-  const prepared = assistantPreparedGroupMessageSchema.safeParse({
-    chat_group_id: groupResolution.chatGroupId,
-    group_name: groupResolution.groupName,
-    message_status: groupResolution.messageStatus,
-    body: parsedDraft.data.body,
-  });
-
-  if (!prepared.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: prepared.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: draftWithResolvedGroup,
-      },
-    };
-  }
-
-  const { data: org, error: orgError } = await sb
-    .from("organizations")
-    .select("slug")
-    .eq("id", ctx.orgId)
-    .maybeSingle();
-
-  if (orgError) {
-    aiLog("warn", "ai-tools", "prepare_group_message org lookup failed", logContext, {
-      error: getSafeErrorMessage(orgError),
-    });
-    return toolError("Failed to load organization context");
-  }
-
-  const pendingPayload: SendGroupChatMessagePendingPayload = {
-    ...prepared.data,
-    orgSlug: typeof org?.slug === "string" ? org.slug : null,
-  };
-  const created = await createOrRevisePendingAction(sb, {
-    organizationId: ctx.orgId,
-    userId: ctx.userId,
-    threadId: ctx.threadId,
-    actionType: "send_group_chat_message",
-    payload: pendingPayload,
-    activeActionId: ctx.activePendingActionId,
-  });
-  if ("failed" in created) return pendingActionFailureToToolError(created.reason);
-  return {
-    kind: "ok",
-    data: {
-      state: "needs_confirmation",
-      draft: draftWithResolvedGroup,
-      pending_action: buildPendingActionField(created, pendingPayload),
-    },
-  };
-}
-
-async function prepareEvent(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareEventSchema>
-): Promise<ToolExecutionResult> {
-  const logContext = buildLogContext(ctx);
-  if (!ctx.threadId) {
-    return toolError("Event preparation requires a thread context");
-  }
-
-  // Normalize empty strings to undefined (LLM often sends "")
-  const normalized = Object.fromEntries(
-    Object.entries(args).filter(
-      ([, v]) => !(typeof v === "string" && v.trim().length === 0)
-    )
-  ) as typeof args;
-
-  const draftWithDefaults = {
-    ...normalized,
-    event_type: normalized.event_type ?? "general",
-    is_philanthropy: normalized.is_philanthropy ?? normalized.event_type === "philanthropy",
-  };
-
-  const parsedDraft = assistantEventDraftSchema.safeParse(draftWithDefaults);
-  if (!parsedDraft.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: parsedDraft.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: draftWithDefaults,
-      },
-    };
-  }
-
-  const missingFields = REQUIRED_PREPARED_EVENT_FIELDS.filter((field) => {
-    const value = parsedDraft.data[field];
-    return typeof value !== "string" || value.trim().length === 0;
-  });
-
-  if (missingFields.length > 0) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: missingFields,
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const prepared = assistantPreparedEventSchema.safeParse(parsedDraft.data);
-  if (!prepared.success) {
-    return {
-      kind: "ok",
-      data: {
-        state: "missing_fields",
-        missing_fields: prepared.error.issues.map((issue) => issue.path.join(".") || "body"),
-        draft: parsedDraft.data,
-      },
-    };
-  }
-
-  const { data: org, error: orgError } = await sb
-    .from("organizations")
-    .select("slug")
-    .eq("id", ctx.orgId)
-    .maybeSingle();
-
-  if (orgError) {
-    aiLog("warn", "ai-tools", "prepare_event org lookup failed", logContext, {
-      error: getSafeErrorMessage(orgError),
-    });
-    return toolError("Failed to load organization context");
-  }
-
-  const pendingPayload: CreateEventPendingPayload = {
-    ...prepared.data,
-    orgSlug: typeof org?.slug === "string" ? org.slug : null,
-  };
-  const created = await createOrRevisePendingAction(sb, {
-    organizationId: ctx.orgId,
-    userId: ctx.userId,
-    threadId: ctx.threadId,
-    actionType: "create_event",
-    payload: pendingPayload,
-    activeActionId: ctx.activePendingActionId,
-  });
-  if ("failed" in created) return pendingActionFailureToToolError(created.reason);
-  return {
-    kind: "ok",
-    data: {
-      state: "needs_confirmation",
-      draft: prepared.data,
-      pending_action: buildPendingActionField(created, pendingPayload),
-    },
-  };
-}
-
-async function prepareEventsBatch(
-  sb: SB,
-  ctx: ToolExecutionContext,
-  args: z.infer<typeof prepareEventsBatchSchema>
-): Promise<ToolExecutionResult> {
-  const logContext = buildLogContext(ctx);
-  if (!ctx.threadId) {
-    return toolError("Event preparation requires a thread context");
-  }
-
-  // Look up org slug once for all events
-  const { data: org, error: orgError } = await sb
-    .from("organizations")
-    .select("slug")
-    .eq("id", ctx.orgId)
-    .maybeSingle();
-
-  if (orgError) {
-    aiLog("warn", "ai-tools", "prepare_events_batch org lookup failed", logContext, {
-      error: getSafeErrorMessage(orgError),
-    });
-    return toolError("Failed to load organization context");
-  }
-
-  const orgSlug = typeof org?.slug === "string" ? org.slug : null;
-  return {
-    kind: "ok",
-    data: await buildPendingEventBatchFromDrafts(
-      sb,
-      ctx,
-      args.events,
-      logContext,
-      orgSlug
-    ),
   };
 }
 
@@ -2321,389 +814,6 @@ async function createSignedScheduleUploadUrl(
   return signedUrl;
 }
 
-async function getOrgStats(
-  sb: SB,
-  orgId: string,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const [members, alumni, parents, upcomingEvents, donations] = await Promise.all([
-    safeToolCount(logContext, () =>
-      sb
-        .from("members")
-        .select("*", { count: "estimated", head: true })
-        .eq("organization_id", orgId)
-        .is("deleted_at", null)
-        .eq("status", "active")
-    ),
-    safeToolCount(logContext, () =>
-      sb
-        .from("alumni")
-        .select("*", { count: "estimated", head: true })
-        .eq("organization_id", orgId)
-        .is("deleted_at", null)
-    ),
-    safeToolCount(logContext, () =>
-      sb
-        .from("parents")
-        .select("*", { count: "estimated", head: true })
-        .eq("organization_id", orgId)
-        .is("deleted_at", null)
-    ),
-    safeToolCount(logContext, () =>
-      sb
-        .from("events")
-        .select("*", { count: "estimated", head: true })
-        .eq("organization_id", orgId)
-        .is("deleted_at", null)
-        .gte("start_date", new Date().toISOString())
-    ),
-    safeToolQuery(logContext, () =>
-      sb
-        .from("organization_donation_stats")
-        .select("total_amount_cents, donation_count, last_donation_at")
-        .eq("organization_id", orgId)
-        .maybeSingle()
-    ),
-  ]);
-
-  if (!members.ok || !alumni.ok || !parents.ok || !upcomingEvents.ok || donations.kind !== "ok") {
-    return toolError("Query failed");
-  }
-
-  return {
-    kind: "ok",
-    data: {
-      active_members: members.count,
-      alumni: alumni.count,
-      parents: parents.count,
-      upcoming_events: upcomingEvents.count,
-      donations: donations.data,
-    },
-  };
-}
-
-function defaultDonationAnalyticsBucket(windowDays: number): "day" | "week" | "month" {
-  if (windowDays <= 31) return "day";
-  if (windowDays <= 180) return "week";
-  return "month";
-}
-
-async function getDonationAnalytics(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof getDonationAnalyticsSchema>,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const windowDays = args.window_days ?? 90;
-  const bucket = args.bucket ?? defaultDonationAnalyticsBucket(windowDays);
-  const topPurposesLimit = args.top_purposes_limit ?? 5;
-
-  return safeToolQuery(logContext, async () => {
-    const { data, error } = await sb.rpc("get_donation_analytics", {
-      p_org_id: orgId,
-      p_window_days: windowDays,
-      p_bucket: bucket,
-      p_top_purposes_limit: topPurposesLimit,
-    });
-
-    return { data, error };
-  });
-}
-
-async function findNavigationTargets(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof findNavigationTargetsSchema>,
-  ctx: ToolExecutionContext,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  return safeToolQuery(logContext, async () => {
-    const [orgResult, subscriptionResult] = await Promise.all([
-      sb
-        .from("organizations")
-        .select("slug, nav_config")
-        .eq("id", orgId)
-        .maybeSingle(),
-      sb.rpc("get_subscription_status", { p_org_id: orgId }),
-    ]);
-
-    const { data: org, error } = orgResult;
-    if (error || !org?.slug) {
-      return { data: null, error: error ?? new Error("Organization not found") };
-    }
-
-    const { data: subscriptionRows, error: subscriptionError } = subscriptionResult;
-
-    const subscription = subscriptionError
-      ? null
-      : Array.isArray(subscriptionRows)
-        ? subscriptionRows[0]
-        : null;
-    const hasAlumniAccess =
-      subscription?.status === "enterprise_managed" ||
-      (subscription?.alumni_bucket != null && subscription.alumni_bucket !== "none");
-    const hasParentsAccess =
-      subscription?.status === "enterprise_managed" ||
-      (subscription?.parents_bucket != null && subscription.parents_bucket !== "none");
-
-    return {
-      data: searchNavigationTargets({
-        query: args.query,
-        orgSlug: org.slug,
-        navConfig:
-          org.nav_config && typeof org.nav_config === "object" && !Array.isArray(org.nav_config)
-            ? (org.nav_config as NavConfig)
-            : null,
-        role: ctx.authorization.kind === "preverified_role" ? ctx.authorization.role : "admin",
-        hasAlumniAccess,
-        hasParentsAccess,
-        limit: args.limit,
-      }),
-      error: null,
-    };
-  });
-}
-
-async function runSuggestConnections(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof suggestConnectionsSchema>,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const {
-    suggestConnections,
-    SuggestConnectionsLookupError,
-  } = await import("@/lib/falkordb/suggestions");
-
-  try {
-    const data = await suggestConnections({
-      orgId,
-      serviceSupabase: sb,
-      args,
-    });
-
-    return {
-      kind: "ok",
-      data,
-    };
-  } catch (error) {
-    if (error instanceof SuggestConnectionsLookupError) {
-      return toolError(error.message);
-    }
-
-    if (isStageTimeoutError(error)) {
-      throw error;
-    }
-
-    aiLog("warn", "ai-tools", "suggest_connections failed", logContext, {
-      error: getSafeErrorMessage(error),
-    });
-    return toolError("Unexpected error");
-  }
-}
-
-async function runSuggestMentors(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof suggestMentorsSchema>,
-  ctx: ToolExecutionContext,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  // Amendment D: admin-only in v1
-  if (ctx.authorization.kind !== "preverified_admin") {
-    aiLog("info", "ai-tools", "suggest_mentors unauthorized", logContext, {
-      auth_decision: "unauthorized",
-    });
-    return {
-      kind: "ok",
-      data: { state: "unauthorized", message: "Mentor suggestions are currently admin-only." },
-    };
-  }
-
-  const { suggestMentors } = await import("@/lib/mentorship/ai-suggestions");
-
-  try {
-    const data = await suggestMentors(sb as unknown as import("@supabase/supabase-js").SupabaseClient<import("@/types/database").Database>, orgId, {
-      menteeUserId: args.mentee_id,
-      menteeQuery: args.mentee_query,
-      focusAreas: args.focus_areas,
-      limit: args.limit,
-    });
-
-    aiLog("info", "ai-tools", "suggest_mentors completed", logContext, {
-      auth_decision: "allowed",
-      state: data.state,
-      candidate_count: data.suggestions.length,
-      top_reason_codes: data.suggestions[0]?.reasons.map((r) => r.code) ?? [],
-    });
-
-    return { kind: "ok", data };
-  } catch (error) {
-    if (isStageTimeoutError(error)) throw error;
-
-    aiLog("warn", "ai-tools", "suggest_mentors failed", logContext, {
-      error: getSafeErrorMessage(error),
-    });
-    return toolError("Unexpected error");
-  }
-}
-
-async function runListAvailableMentors(
-  sb: SB,
-  orgId: string,
-  args: z.infer<typeof listAvailableMentorsSchema>,
-  logContext: AiLogContext
-): Promise<ToolExecutionResult> {
-  const limit = args.limit ?? 5;
-
-  try {
-    const { data: mentorProfiles, error } = await sb
-      .from("mentor_profiles")
-      .select(
-        "user_id, topics, sports, positions, max_mentees, current_mentee_count, accepting_new, is_active"
-      )
-      .eq("organization_id", orgId)
-      .eq("is_active", true)
-      .eq("accepting_new", true);
-
-    if (error) {
-      aiLog("warn", "ai-tools", "list_available_mentors query failed", logContext, {
-        error: getSafeErrorMessage(error),
-      });
-      return toolError("Query failed");
-    }
-
-    const eligibleMentors = (Array.isArray(mentorProfiles) ? mentorProfiles : [])
-      .map((profile) => ({
-        user_id: typeof profile.user_id === "string" ? profile.user_id : null,
-        topics: Array.isArray(profile.topics)
-          ? profile.topics.filter((value: unknown): value is string => typeof value === "string")
-          : [],
-        sports: Array.isArray(profile.sports)
-          ? profile.sports.filter((value: unknown): value is string => typeof value === "string")
-          : [],
-        positions: Array.isArray(profile.positions)
-          ? profile.positions.filter((value: unknown): value is string => typeof value === "string")
-          : [],
-        max_mentees: typeof profile.max_mentees === "number" ? profile.max_mentees : 3,
-        current_mentee_count:
-          typeof profile.current_mentee_count === "number"
-            ? profile.current_mentee_count
-            : 0,
-      }))
-      .filter(
-        (profile): profile is {
-          user_id: string;
-          topics: string[];
-          sports: string[];
-          positions: string[];
-          max_mentees: number;
-          current_mentee_count: number;
-        } =>
-          Boolean(profile.user_id) &&
-          profile.max_mentees > profile.current_mentee_count
-      );
-
-    if (eligibleMentors.length === 0) {
-      return {
-        kind: "ok",
-        data: { state: "no_results", total_available: 0, mentors: [] },
-      };
-    }
-
-    const mentorUserIds = eligibleMentors.map((mentor) => mentor.user_id);
-    const [{ data: mentorUsers, error: usersError }, { data: alumniRows, error: alumniError }] =
-      await Promise.all([
-        sb.from("users").select("id, name, email").in("id", mentorUserIds),
-        sb
-          .from("alumni")
-          .select("user_id, job_title, current_company")
-          .eq("organization_id", orgId)
-          .in("user_id", mentorUserIds),
-      ]);
-
-    if (usersError || alumniError) {
-      aiLog("warn", "ai-tools", "list_available_mentors enrichment failed", logContext, {
-        error: getSafeErrorMessage(usersError ?? alumniError),
-      });
-      return toolError("Query failed");
-    }
-
-    const userLookup = new Map<string, { name: string | null; email: string | null }>();
-    for (const row of Array.isArray(mentorUsers) ? mentorUsers : []) {
-      if (typeof row.id !== "string") continue;
-      userLookup.set(row.id, {
-        name: typeof row.name === "string" ? row.name : null,
-        email: typeof row.email === "string" ? row.email : null,
-      });
-    }
-
-    const alumniLookup = new Map<
-      string,
-      { job_title: string | null; current_company: string | null }
-    >();
-    for (const row of Array.isArray(alumniRows) ? alumniRows : []) {
-      if (typeof row.user_id !== "string") continue;
-      alumniLookup.set(row.user_id, {
-        job_title: typeof row.job_title === "string" ? row.job_title : null,
-        current_company:
-          typeof row.current_company === "string" ? row.current_company : null,
-      });
-    }
-
-    const mentors = eligibleMentors
-      .sort((left, right) => {
-        const leftOpenSlots = left.max_mentees - left.current_mentee_count;
-        const rightOpenSlots = right.max_mentees - right.current_mentee_count;
-        if (rightOpenSlots !== leftOpenSlots) return rightOpenSlots - leftOpenSlots;
-        return left.current_mentee_count - right.current_mentee_count;
-      })
-      .slice(0, limit)
-      .map((mentor) => {
-        const user = userLookup.get(mentor.user_id);
-        const alumni = alumniLookup.get(mentor.user_id);
-        const subtitle = [alumni?.job_title, alumni?.current_company]
-          .filter((value): value is string => typeof value === "string" && value.length > 0)
-          .join(" at ");
-
-        return {
-          mentor: {
-            user_id: mentor.user_id,
-            name: user?.name ?? user?.email ?? "Mentor",
-            subtitle: subtitle || null,
-          },
-          open_slots: mentor.max_mentees - mentor.current_mentee_count,
-          current_mentee_count: mentor.current_mentee_count,
-          max_mentees: mentor.max_mentees,
-          topics: mentor.topics.slice(0, 3),
-          sports: mentor.sports.slice(0, 3),
-          positions: mentor.positions.slice(0, 3),
-        };
-      });
-
-    aiLog("info", "ai-tools", "list_available_mentors completed", logContext, {
-      total_available: eligibleMentors.length,
-      returned_count: mentors.length,
-    });
-
-    return {
-      kind: "ok",
-      data: {
-        state: "resolved",
-        total_available: eligibleMentors.length,
-        mentors,
-      },
-    };
-  } catch (error) {
-    if (isStageTimeoutError(error)) throw error;
-
-    aiLog("warn", "ai-tools", "list_available_mentors failed", logContext, {
-      error: getSafeErrorMessage(error),
-    });
-    return toolError("Unexpected error");
-  }
-}
-
 async function verifyExecutorAccess(
   ctx: ToolExecutionContext
 ): Promise<{ kind: "allowed" } | Extract<ToolExecutionResult, { kind: "forbidden" | "auth_error" }>> {
@@ -2859,37 +969,6 @@ export async function executeToolCall(
         return dispatchToolModule(toolName, args, { ctx, sb, logContext });
       }
       switch (toolName) {
-        case "list_donations":
-          return listDonations(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof listDonationsSchema>,
-            logContext
-          );
-        case "list_parents":
-          return listParents(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof listParentsSchema>,
-            logContext
-          );
-        case "list_philanthropy_events":
-          return listPhilanthropyEvents(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof listPhilanthropyEventsSchema>,
-            logContext
-          );
-        case "list_managed_orgs":
-          return safeToolQuery(logContext, () => listManagedOrgs(sb, ctx.enterpriseId!));
-        case "list_enterprise_audit_events":
-          return safeToolQuery(logContext, () =>
-            listEnterpriseAuditEvents(
-              sb,
-              ctx.enterpriseId!,
-              args as z.infer<typeof listEnterpriseAuditEventsSchema>,
-            )
-          );
         case "prepare_enterprise_invite":
           return prepareEnterpriseInvite(
             sb,
@@ -2902,54 +981,6 @@ export async function executeToolCall(
             ctx,
             args as z.infer<typeof revokeEnterpriseInviteSchema>,
           );
-        case "prepare_announcement":
-          return prepareAnnouncement(
-            sb,
-            ctx,
-            args as z.infer<typeof prepareAnnouncementSchema>
-          );
-        case "prepare_job_posting":
-          return prepareJobPosting(
-            sb,
-            ctx,
-            args as z.infer<typeof prepareJobPostingSchema>
-          );
-        case "prepare_chat_message":
-          return prepareChatMessage(
-            sb,
-            ctx,
-            args as z.infer<typeof prepareChatMessageSchema>
-          );
-        case "prepare_group_message":
-          return prepareGroupMessage(
-            sb,
-            ctx,
-            args as z.infer<typeof prepareGroupMessageSchema>
-          );
-        case "prepare_discussion_reply":
-          return prepareDiscussionReply(
-            sb,
-            ctx,
-            args as z.infer<typeof prepareDiscussionReplySchema>
-          );
-        case "prepare_discussion_thread":
-          return prepareDiscussionThread(
-            sb,
-            ctx,
-            args as z.infer<typeof prepareDiscussionThreadSchema>
-          );
-        case "prepare_event":
-          return prepareEvent(
-            sb,
-            ctx,
-            args as z.infer<typeof prepareEventSchema>
-          );
-        case "prepare_events_batch":
-          return prepareEventsBatch(
-            sb,
-            ctx,
-            args as z.infer<typeof prepareEventsBatchSchema>
-          );
         case "scrape_schedule_website":
           return scrapeScheduleWebsite(
             sb,
@@ -2961,51 +992,6 @@ export async function executeToolCall(
             sb,
             ctx,
             args as z.infer<typeof extractSchedulePdfSchema>
-          );
-        case "get_org_stats":
-          return getOrgStats(sb, ctx.orgId, logContext);
-        case "get_donation_analytics":
-          return getDonationAnalytics(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof getDonationAnalyticsSchema>,
-            logContext
-          );
-        case "get_enterprise_stats":
-          return safeToolQuery(logContext, () => getEnterpriseStats(sb, ctx.enterpriseId!));
-        case "get_enterprise_quota":
-          return safeToolQuery(logContext, () => getEnterpriseQuota(sb, ctx.enterpriseId!));
-        case "get_enterprise_org_capacity":
-          return safeToolQuery(logContext, () => getEnterpriseOrgCapacity(sb, ctx.enterpriseId!));
-        case "suggest_connections":
-          return runSuggestConnections(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof suggestConnectionsSchema>,
-            logContext
-          );
-        case "suggest_mentors":
-          return runSuggestMentors(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof suggestMentorsSchema>,
-            ctx,
-            logContext
-          );
-        case "list_available_mentors":
-          return runListAvailableMentors(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof listAvailableMentorsSchema>,
-            logContext
-          );
-        case "find_navigation_targets":
-          return findNavigationTargets(
-            sb,
-            ctx.orgId,
-            args as z.infer<typeof findNavigationTargetsSchema>,
-            ctx,
-            logContext
           );
         default:
           // Registered tools dispatched above; remaining cases are exhaustive.
